update 27/03/2023 10:00

+++ lib/models/admin_page_class.dart
@@ -14,7 +14,7 @@ class Site {
   String siteName;
   int monthlyBudget;
   int additionalBudget;
-  int siteArea;
+  double siteArea;
   bool isExpanded;
 
   Map<String, String> toJson() => {
